Olivier demo prep (#73)
* sign in error handling

* error handling for auth/verify endpoint

* adding awaits to asunc functions

* Update UserInterceptor to create new user in database on first sign in

* update default database name

* update app and application table

* include reviews relation in application queries

* Refactor review modal and Application Table

* refactor types

* add Reviews view
