+++ apps/frontend/src/components/ApplicationTables/columns.ts
@@ -1,0 +1,42 @@
+export const applicationColumns = [
+  {
+    field: 'firstName',
+    headerName: 'First Name',
+    width: 150,
+  },
+  {
+    field: 'lastName',
+    headerName: 'Last Name',
+    width: 150,
+  },
+  {
+    field: 'stage',
+    headerName: 'Stage',
+    width: 125,
+  },
+  {
+    field: 'step',
+    headerName: 'Status',
+    width: 125,
+  },
+  {
+    field: 'position',
+    headerName: 'Position',
+    width: 150,
+  },
+  {
+    field: 'createdAt',
+    headerName: 'Date',
+    width: 150,
+  },
+  {
+    field: 'meanRatingAllStages',
+    headerName: 'Rating All Stages',
+    width: 150,
+  },
+  {
+    field: 'meanRatingSingleStages',
+    headerName: 'Rating Single Stage',
+    width: 150,
+  },
+];
