+++ apps/frontend/src/components/ApplicationTables/index.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useState, useRef } from 'react';
 import { DataGrid, GridRowSelectionModel } from '@mui/x-data-grid';
 import {
   Container,
@@ -8,70 +9,14 @@ import {
   ListItemText,
   ListItemIcon,
   Button,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  TextField,
-  Rating,
 } from '@mui/material';
-import { useEffect, useState, useRef } from 'react';
-import apiClient from '@api/apiClient';
 import { DoneOutline } from '@mui/icons-material';
-export enum ApplicationStage {
-  RESUME = 'RESUME',
-  INTERVIEW = 'INTERVIEW',
-  ACCEPTED = 'ACCEPTED',
-  REJECTED = 'REJECTED',
-  TECHNICAL_CHALLENGE = 'TECHNICAL_CHALLENGE',
-  PM_CHALLENGE = 'PM_CHALLENGE',
-}
 
-export enum ApplicationStep {
-  SUBMITTED = 'SUBMITTED',
-  REVIEWED = 'REVIEWED',
-}
+import { ApplicationRow, Application, Semester } from '../types';
+import apiClient from '@api/apiClient';
+import { applicationColumns } from './columns';
+import { ReviewModal } from './reviewModal';
 
-enum Position {
-  DEVELOPER = 'DEVELOPER',
-  PM = 'PRODUCT_MANAGER',
-  DESIGNER = 'DESIGNER',
-}
-
-export type applicationRow = {
-  id: number;
-  userId: number;
-  firstName: string;
-  lastName: string;
-  stage: ApplicationStage;
-  step: ApplicationStep;
-  position: Position;
-  createdAt: string;
-  meanRatingAllStages: number;
-  meanRatingSingleStages: number;
-};
-
-type Response = {
-  question: string;
-  answer: string;
-};
-
-enum Semester {
-  FALL = 'FALL',
-  SPRING = 'SPRING',
-}
-
-export type Application = {
-  id: number;
-  createdAt: Date;
-  year: number;
-  semester: Semester;
-  position: Position;
-  stage: ApplicationStage;
-  step: ApplicationStep;
-  response: Response[];
-  numApps: number;
-};
 const TODAY = new Date();
 
 const getCurrentSemester = (): Semester => {
@@ -89,47 +34,23 @@ const getCurrentYear = (): number => {
 export function ApplicationTable() {
   const isPageRendered = useRef<boolean>(false);
 
-  const [data, setData] = useState<applicationRow[]>([]);
+  // TODO switch to use code grant flow
+  // TODO automatically redirect to login page if not logged in
+  // TODO implement auto token refresh
+  const [data, setData] = useState<ApplicationRow[]>([]);
   const [fullName, setFullName] = useState<string>('');
   const [accessToken, setAccessToken] = useState<string>('');
   const [rowSelection, setRowSelection] = useState<GridRowSelectionModel>([]);
-  const [selectedUser, setSelectedUser] = useState<applicationRow | null>(null);
+  const [selectedUserRow, setSelectedUserRow] = useState<ApplicationRow | null>(
+    null,
+  );
   const [selectedApplication, setSelectedApplication] =
     useState<Application | null>(null);
 
   const [openReviewModal, setOpenReviewModal] = useState(false);
-  const [reviewComment, setReviewComment] = useState('');
-  const [reviewRating, setReviewRating] = useState<number>(0);
 
   const handleOpenReviewModal = () => {
     setOpenReviewModal(true);
-  };
-
-  const handleCloseReviewModal = () => {
-    setOpenReviewModal(false);
-    setReviewComment('');
-  };
-  const stageToSubmit = selectedApplication?.stage || ApplicationStage.ACCEPTED;
-
-  const handleReviewSubmit = async () => {
-    if (!selectedUser || reviewRating === 0 || !reviewComment) {
-      alert('Please select a user, provide a rating, and add a comment.');
-      return;
-    }
-
-    try {
-      await apiClient.submitReview(accessToken, {
-        applicantId: selectedUser.userId,
-        stage: stageToSubmit,
-        rating: reviewRating,
-        content: reviewComment,
-      });
-      alert('Review submitted successfully!');
-      handleCloseReviewModal();
-    } catch (error) {
-      console.error('Error submitting review:', error);
-      alert('Failed to submit review.');
-    }
   };
 
   const fetchData = async () => {
@@ -142,11 +63,6 @@ export function ApplicationTable() {
       setData(data);
     }
   };
-
-  // const getApplication = async (userId: number) => {
-  //   const application = await apiClient.getApplication(accessToken, userId);
-  //   setSelectedApplication(application);
-  // };
 
   const getApplication = async (userId: number) => {
     try {
@@ -183,7 +99,7 @@ export function ApplicationTable() {
 
   useEffect(() => {
     if (rowSelection.length > 0) {
-      setSelectedUser(data[rowSelection[0] as number]);
+      setSelectedUserRow(data[rowSelection[0] as number]);
     }
   }, [rowSelection, data]);
 
@@ -200,48 +116,7 @@ export function ApplicationTable() {
       </Typography>
       <DataGrid
         rows={data}
-        columns={[
-          {
-            field: 'firstName',
-            headerName: 'First Name',
-            width: 150,
-          },
-          {
-            field: 'lastName',
-            headerName: 'Last Name',
-            width: 150,
-          },
-          {
-            field: 'stage',
-            headerName: 'Stage',
-            width: 125,
-          },
-          {
-            field: 'step',
-            headerName: 'Status',
-            width: 125,
-          },
-          {
-            field: 'position',
-            headerName: 'Position',
-            width: 150,
-          },
-          {
-            field: 'createdAt',
-            headerName: 'Date',
-            width: 150,
-          },
-          {
-            field: 'meanRatingAllStages',
-            headerName: 'Rating All Stages',
-            width: 150,
-          },
-          {
-            field: 'meanRatingSingleStages',
-            headerName: 'Rating Single Stage',
-            width: 150,
-          },
-        ]}
+        columns={applicationColumns}
         initialState={{
           pagination: {
             paginationModel: { page: 0, pageSize: 5 },
@@ -256,8 +131,8 @@ export function ApplicationTable() {
       />
 
       <Typography variant="h6" mt={3}>
-        {selectedUser
-          ? `Selected Applicant: ${selectedUser.firstName} ${selectedUser.lastName}`
+        {selectedUserRow
+          ? `Selected Applicant: ${selectedUserRow.firstName} ${selectedUserRow.lastName}`
           : 'No Applicant Selected'}
       </Typography>
       {selectedApplication ? (
@@ -301,14 +176,25 @@ export function ApplicationTable() {
               </ListItem>
             ))}
           </List>
-          <Stack
-            direction="row"
-            spacing={1}
-            sx={{
-              alignItems: 'center',
-            }}
-          >
-            <Typography variant="body1">Reviews: None</Typography>
+          <Stack>
+            <Typography variant="body1">
+              Reviews:
+              {selectedApplication.reviews.map((review, index) => {
+                return (
+                  <Stack key={index} direction="row" spacing={1}>
+                    <Typography variant="body1">
+                      stage: {review.stage}
+                    </Typography>
+                    <Typography variant="body1">
+                      rating: {review.rating}
+                    </Typography>
+                    <Typography variant="body1">
+                      comment: {review.content}
+                    </Typography>
+                  </Stack>
+                );
+              })}
+            </Typography>
             <Button
               variant="contained"
               size="small"
@@ -317,37 +203,13 @@ export function ApplicationTable() {
               Start Review
             </Button>
           </Stack>
-          <Dialog open={openReviewModal} onClose={handleCloseReviewModal}>
-            <DialogTitle>Write Review</DialogTitle>
-            <DialogContent>
-              <Stack direction="row" alignItems="center" spacing={2} mb={2}>
-                <Typography variant="body1">Rating:</Typography>
-                <Rating
-                  name="review-rating"
-                  value={reviewRating}
-                  onChange={(_, value) => setReviewRating(value || 0)}
-                  precision={1}
-                />
-              </Stack>
-              <TextField
-                autoFocus
-                margin="dense"
-                id="review"
-                label="Review Comments"
-                type="text"
-                fullWidth
-                multiline
-                rows={4}
-                variant="outlined"
-                value={reviewComment}
-                onChange={(e) => setReviewComment(e.target.value)}
-              />
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={handleCloseReviewModal}>Cancel</Button>
-              <Button onClick={handleReviewSubmit}>Submit Review</Button>
-            </DialogActions>
-          </Dialog>
+          <ReviewModal
+            open={openReviewModal}
+            setOpen={setOpenReviewModal}
+            selectedUserRow={selectedUserRow}
+            selectedApplication={selectedApplication}
+            accessToken={accessToken}
+          />
         </>
       ) : null}
     </Container>
