+++ apps/frontend/src/containers/root.tsx
@@ -1,4 +1,4 @@
-import { ApplicationTable } from '@components/ApplicationTables.jsx';
+import { ApplicationTable } from '@components/ApplicationTables';
 
 const Root: React.FC = () => {
   return <ApplicationTable />;
